api/requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==2.3.3
Flask-MySQL==1.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyMySQL==1.1.0
PyYAML==6.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
